Split automation dispatch event handling
